dermascan/backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
tqdm>=4.61.2
requests>=2.26.0
